requirements.txt
flask
pytesseract
pillow
PyMuPDF
